# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: py-payments-ledger
  version: 0.1.0
servers:
- url: http://localhost:8000
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        correlation_id:
          type: string
      required:
      - title
      - status
      - detail
      - instance
      - correlation_id
    TokenRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        tenantId:
          type:
          - string
          - 'null'
      required:
      - email
      - password
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
      required:
      - access_token
      - token_type
      - expires_in
...
paths:
  /v1/auth/token:
    post:
      tags:
      - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1/me:
    get:
      tags:
      - auth
      security:
      - BearerAuth: []
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...